requirements.txt
matplotlib>=3.8.0
numpy>=1.26.0
pillow>=11.3.0
streamlit>=1.48.0
